+++ config/core.json
@@ -8,7 +8,7 @@
 		"marketing": true,
 		"navigation": false,
 		"onboarding": true,
-		"rule-based-inbox-notes": false,
+		"remote-inbox-notifications": false,
 	  	"shipping-label-banner": true,
 		"store-alerts": true,
 		"wcpay": true,
