+++ config/development.json
@@ -8,7 +8,7 @@
 		"marketing": true,
 		"navigation": true,
 		"onboarding": true,
-		"rule-based-inbox-notes": true,
+		"remote-inbox-notifications": true,
 		"shipping-label-banner": true,
 		"store-alerts": true,
 		"wcpay": true,
