+++ src/Events.php
@@ -19,7 +19,7 @@ use \Automattic\WooCommerce\Admin\Notes\WC_Admin_Notes_Onboarding_Payments;
 use \Automattic\WooCommerce\Admin\Notes\WC_Admin_Notes_Personalize_Store;
 use \Automattic\WooCommerce\Admin\Notes\WC_Admin_Notes_WooCommerce_Payments;
 use \Automattic\WooCommerce\Admin\Notes\WC_Admin_Notes_Marketing;
-use \Automattic\WooCommerce\Admin\Rinds\RindsSpecsDataSourcePoller;
+use \Automattic\WooCommerce\Admin\RemoteInboxNotifications\DataSourcePoller;
 
 /**
  * WC_Admin_Events Class.
@@ -73,6 +73,6 @@ class Events {
 		WC_Admin_Notes_WooCommerce_Payments::possibly_add_note();
 		WC_Admin_Notes_Marketing::possibly_add_note();
 		WC_Admin_Notes_Giving_Feedback_Notes::possibly_add_note();
-		RindsSpecsDataSourcePoller::poll_data_sources();
+		DataSourcePoller::poll_data_sources();
 	}
 }
