+++ src/RemoteInboxNotifications/PluginsActivatedRuleProcessor.php
@@ -5,7 +5,7 @@
  * @package WooCommerce Admin/Classes;
  */
 
-namespace Automattic\WooCommerce\Admin\Rinds;
+namespace Automattic\WooCommerce\Admin\RemoteInboxNotifications;
 
 defined( 'ABSPATH' ) || exit;
