+++ tests/remote-inbox-notifications/failing-rule-evaluator.php
@@ -2,7 +2,7 @@
 /**
  * FailingRuleEvaluator
  *
- * @package WooCommerce\Tests\Rinds
+ * @package WooCommerce\Tests\RemoteInboxNotifications
  */
 
 /**
