+++ tests/remote-inbox-notifications/plugins-activated-rule-processor.php
@@ -2,16 +2,16 @@
 /**
  * Plugins activated rule processor tests.
  *
- * @package WooCommerce\Tests\Rinds
+ * @package WooCommerce\Tests\RemoteInboxNotifications
  */
 
-use Automattic\WooCommerce\Admin\Rinds\PluginsActivatedRuleProcessor;
+use Automattic\WooCommerce\Admin\RemoteInboxNotifications\PluginsActivatedRuleProcessor;
 use Automattic\WooCommerce\Admin\PluginsProvider\PluginsProviderInterface;
 
 /**
- * class WC_Tests_Rinds_PluginsActivatedRuleProcessor
+ * class WC_Tests_RemoteInboxNotifications_PluginsActivatedRuleProcessor
  */
-class WC_Tests_Rinds_PluginsActivatedRuleProcessor extends WC_Unit_Test_Case {
+class WC_Tests_RemoteInboxNotifications_PluginsActivatedRuleProcessor extends WC_Unit_Test_Case {
 	/**
 	 * Tests that the processor does not pass a plugins_activated rule with
 	 * no plugins to verify.
